correction com.fazecast.jSerialComm.SerialPort

+++ arduino/libraries/AT24C32/I2C_EEPROM.h
@@ -26,8 +26,8 @@ class I2C_EEPROM {
 public:
   uint8_t read(uint8_t device, uint16_t address);
   void write(uint8_t device, uint16_t address, uint8_t data);
-  void write_buffer(uint8_t device, uint16_t address, uint8_t* data, uint8_t length);
-  void read_buffer(uint8_t device, uint16_t address, uint8_t* data, uint8_t length);
+  void write_buffer(uint8_t device, uint16_t address, uint8_t* data, uint16_t length);
+  void read_buffer(uint8_t device, uint16_t address, uint8_t* data, uint16_t length);
 private:
   void beginTransmission(uint8_t device, uint16_t address);
 };
